# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "servers": [
    {
      "url": "https://vodmasterdata.herokuapp.com//"
    }
  ],
  "paths": {
    "/api/category/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/video/": {
      "post": {
        "tags": [
          "Videos"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newvideo",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "The Lord of the Rings: The Fellowship of the Ring"
                },
                "synopsis": {
                  "type": "string",
                  "example": "A meek Hobbit from the Shire and eight companions set out on a journey to destroy the powerful One Ring and save Middle-earth from the Dark Lord Sauron."
                },
                "director": {
                  "type": "string",
                  "example": "Peter Jackson"
                },
                "cast": {
                  "type": "array",
                  "example": [
                    "Elijah Wood",
                    "Ian McKellen",
                    "Orlando Bloom"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "category": {
                  "type": "string",
                  "example": "6350106165ef3bb9384e0afe"
                },
                "poster": {
                  "type": "string",
                  "example": "https://m.media-amazon.com/images/I/71TZ8BmoZqL._AC_SY879_.jpg"
                },
                "streamURL": {
                  "type": "string",
                  "example": "https://www.youtube.com/watch?v=V75dMMIW2B4"
                }
              },
              "required": [
                "title",
                "synopsis",
                "director",
                "cast",
                "category",
                "poster",
                "streamURL"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Video"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Video"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "example": "any"
                  },
                  "synopsis": {
                    "example": "any"
                  },
                  "director": {
                    "example": "any"
                  },
                  "cast": {
                    "example": "any"
                  },
                  "category": {
                    "example": "any"
                  },
                  "poster": {
                    "example": "any"
                  },
                  "streamURL": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Videos"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Video"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Video"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/video/{id}": {
      "get": {
        "tags": [
          "Videos"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Videos"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/video/{id}/rate": {
      "put": {
        "tags": [
          "Videos"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Video",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "rate": {
                  "type": "number",
                  "example": 3
                }
              },
              "required": [
                "rate"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rate": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/signup": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "newuser",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Alice"
                },
                "email": {
                  "type": "string",
                  "example": "user@fake-mail.com"
                },
                "password": {
                  "type": "string",
                  "example": "password1"
                }
              },
              "required": [
                "name",
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "msg": "User user@fake-mail.com successfully signedup.",
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/signin": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user@fake-mail.com"
                },
                "password": {
                  "type": "string",
                  "example": "password1"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "role": "client",
            "token": "eyJhbGciOi.NjIxMDc2MX0.o3CL-Sf8gh6wBj0"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden",
            "msg": "User user@fake-mail.com have no authorization."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Hero": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Spiderman"
          },
          "strength": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "name",
          "strength"
        ],
        "xml": {
          "name": "Hero"
        }
      },
      "Book": {
        "type": "object",
        "properties": {
          "isbn": {
            "type": "string",
            "example": "978-3-16-148410-0"
          },
          "title": {
            "type": "string",
            "example": "Frankenstein"
          },
          "author": {
            "type": "string",
            "example": "Mary Shelley"
          },
          "cover": {
            "type": "string",
            "example": "http://free-books.org/File:Frankenstein_1818_cover.jpg"
          },
          "available": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "isbn",
          "title",
          "author",
          "available"
        ],
        "xml": {
          "name": "Book"
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "634095f0ea85138ff2010b65"
          },
          "name": {
            "type": "string",
            "example": "Coca-Cola"
          },
          "category": {
            "type": "string",
            "example": "drinks"
          },
          "price": {
            "type": "number",
            "example": 1.14
          }
        },
        "required": [
          "name",
          "category"
        ],
        "xml": {
          "name": "Product"
        }
      },
      "Video": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "634095f0ea85138ff2010b65"
          },
          "title": {
            "type": "string",
            "example": "The Lord of the Rings: The Fellowship of the Ring"
          },
          "synopsis": {
            "type": "string",
            "example": "A meek Hobbit from the Shire and eight companions set out on a journey to destroy the powerful One Ring and save Middle-earth from the Dark Lord Sauron."
          },
          "director": {
            "type": "string",
            "example": "Peter Jackson"
          },
          "cast": {
            "type": "array",
            "example": [
              "Elijah Wood",
              "Ian McKellen",
              "Orlando Bloom"
            ],
            "items": {
              "type": "string"
            }
          },
          "category": {
            "type": "string",
            "example": "6350106165ef3bb9384e0afe"
          },
          "poster": {
            "type": "string",
            "example": "https://m.media-amazon.com/images/I/71TZ8BmoZqL._AC_SY879_.jpg"
          },
          "streamURL": {
            "type": "string",
            "example": "https://www.youtube.com/watch?v=V75dMMIW2B4"
          },
          "rate": {
            "type": "array",
            "example": [
              1,
              3,
              5
            ],
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "title",
          "synopsis",
          "director",
          "cast",
          "category",
          "poster",
          "streamURL"
        ],
        "xml": {
          "name": "Video"
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "6350106165ef3bb9384e0afe"
          },
          "name": {
            "type": "string",
            "example": "Romance"
          }
        },
        "required": [
          "name"
        ],
        "xml": {
          "name": "Category"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}